+++ client/pages/Admin/SummaryView/WeekColumn/index.tsx
@@ -1,21 +1,39 @@
-import { Icon } from 'office-ui-fabric-react'
-import React from 'react'
-import { isEmpty } from 'underscore'
+/* eslint-disable tsdoc/syntax */
+import { TooltipHost } from 'office-ui-fabric-react'
+import React, { FunctionComponent, useRef } from 'react'
+import { useTranslation } from 'react-i18next'
 import { IWeekColumnProps } from './types'
+import { useWeekColumn } from './useWeekColumn'
 import styles from './WeekColumn.module.scss'
+import { WeekColumnTooltip } from './WeekColumnTooltip'
 
-export const WeekColumn = ({ periods }: IWeekColumnProps) => {
-  if (isEmpty(periods)) {
+/**
+ * @category SummaryView
+ */
+export const WeekColumn: FunctionComponent<IWeekColumnProps> = (
+  props: IWeekColumnProps
+) => {
+  const { t } = useTranslation()
+  const target = useRef()
+  const hours = useWeekColumn(props)
+  if (hours.total === null) {
     return null
   }
-
-  const hours = periods.reduce((sum, period) => sum + period.hours, 0)
-
   return (
-    <div className={styles.root}>
-      <Icon iconName='CheckboxComposite' className={styles.checkMark} />
-      <span>{hours}h</span>
-    </div>
+    <TooltipHost
+      calloutProps={{
+        calloutMaxWidth: 420,
+        target
+      }}
+      tooltipProps={{
+        onRenderContent: () => <WeekColumnTooltip {...props} hours={hours} />
+      }}>
+      <div className={styles.root}>
+        <span ref={target}>
+          {t('common.hoursShortFormat', { hours: hours.total.toFixed(0) })}
+        </span>
+      </div>
+    </TooltipHost>
   )
 }