+++ client/pages/Admin/SummaryView/index.tsx
@@ -1,12 +1,11 @@
 /* eslint-disable tsdoc/syntax */
 import { List, UserMessage } from 'components'
-import { IColumn, Pivot, PivotItem } from 'office-ui-fabric-react'
+import { IColumn } from 'office-ui-fabric-react'
 import React from 'react'
 import { useTranslation } from 'react-i18next'
 import { isEmpty } from 'underscore'
 import { useSummaryView } from './hooks/useSummaryView'
 import styles from './SummaryView.module.scss'
-import { ISummaryViewScope } from './types'
 import { WeekColumn } from './WeekColumn'
 
 /**
@@ -14,38 +13,30 @@ import { WeekColumn } from './WeekColumn'
  */
 export const SummaryView = (): JSX.Element => {
   const { t } = useTranslation()
-  const { dispatch, loading, scopes, rows, columns } = useSummaryView({
+  const { state, loading, rows, columns } = useSummaryView({
     onColumnRender: (item: any, _index: number, column: IColumn) => (
-      <WeekColumn periods={item[column.fieldName]} />
+      <WeekColumn
+        user={item.user}
+        periods={item[column.fieldName]}
+        projects={state.projects}
+      />
     )
   })
 
   return (
     <div className={styles.root}>
-      <Pivot
-        onLinkClick={(item) =>
-          dispatch({
-            type: 'CHANGE_SCOPE',
-            payload: item.props as ISummaryViewScope
-          })
-        }>
-        {scopes.map((scope) => (
-          <PivotItem key={scope.itemKey} {...scope}>
-            <div className={styles.container}>
-              <List
-                hidden={!loading && isEmpty(rows)}
-                enableShimmer={loading}
-                columns={columns}
-                items={rows}
-              />
-              <UserMessage
-                hidden={!isEmpty(rows) || loading}
-                text={t('admin.noTimeEntriesText')}
-              />
-            </div>
-          </PivotItem>
-        ))}
-      </Pivot>
+      <div className={styles.container}>
+        <List
+          hidden={!loading && isEmpty(rows)}
+          enableShimmer={loading}
+          columns={columns}
+          items={rows}
+        />
+        <UserMessage
+          hidden={!isEmpty(rows) || loading}
+          text={t('admin.noTimeEntriesText')}
+        />
+      </div>
     </div>
   )
 }