+++ client/pages/Admin/SummaryView/reducer.ts
@@ -1,29 +1,39 @@
-import { ISummaryViewScope, ISummaryViewState } from './types'
+/* eslint-disable react-hooks/exhaustive-deps */
+import { QueryResult } from '@apollo/client'
+import { createAction, createReducer } from '@reduxjs/toolkit'
+import { useMemo, useReducer } from 'react'
+import { ISummaryViewState } from './types'
 
-export type SummaryViewAction =
-  | { type: 'DATA_UPDATED'; payload: any }
-  | { type: 'CHANGE_SCOPE'; payload: ISummaryViewScope }
+export const DATA_UPDATED = createAction<{ data: QueryResult<any> }>(
+  'DATA_UPDATED'
+)
 
-export const reducer = (
-  state: ISummaryViewState,
-  action: SummaryViewAction
-): ISummaryViewState => {
-  const newState: ISummaryViewState = { ...state }
-  switch (action.type) {
-    case 'DATA_UPDATED':
-      {
-        if (action.payload) {
-          newState.users = action.payload.users
-          newState.periods = action.payload.periods
-        }
+function createReducer_(initialState: ISummaryViewState) {
+  return createReducer(initialState, {
+    [DATA_UPDATED.type]: (
+      state,
+      { payload }: ReturnType<typeof DATA_UPDATED>
+    ) => {
+      if (payload.data) {
+        state.periods = payload.data['periods']
+        state.users = payload.data['users']
+        state.projects = payload.data['projects']
       }
-      break
+    }
+  })
+}
 
-    case 'CHANGE_SCOPE':
-      {
-        newState.scope = action.payload
-      }
-      break
+/**
+ * Reducer hook for SummaryView
+ *
+ * @returns React.useReducer with parameters
+ */
+export function useSummaryViewReducer() {
+  const initialState = {
+    users: [],
+    periods: [],
+    projects: []
   }
-  return newState
+  const reducer = useMemo(() => createReducer_(initialState), [initialState])
+  return useReducer(reducer, initialState)
 }